Make side-navbar for mobile #23, Background theme color
When the browser is shrunk to mobile view, a menu icon can be clicked to display an angular side-navigation-bar.
In styles.scss, you can change instances of every html element, so I fixed the issue of previous iterations where background colors broke off at the bottom of the page. Now, the background color is the same and consistent for every page as long as they are wrapped in <html> and <body>.