goal->emotion naming, text field, page restart #36
The restart button did not work completely, so I put a page refresh function inside of it. Also, the emotion controller in server still had "goal" in all of the comments and console prints, so those are now changed to "emotion". The text field is updated to be material design.