work on centering videos #36
only method that has worked so far in resizing videos and centering is adding those properties in the html. I have tried modifiying those properties in the css but they are not working, therefore leaving the currently working on html. Resizing still needs to be done as it does not resize appropriately to devices.

remove labels because on the phone view, the labels were too long and looked messy.