Mat icons not found on droplet, temp fix
The material icons are not found on the droplet because of server updates that fixed page refreshing, so now we are going to the hashtag url referencing which refreshes fine on the droplet but at the cost of showing a # on every link. This should fix the material icons not being found.